refactor(all) : ladder/common -> ladder/attribute 로 package 명 변경
