+++ src/main/java/domain/ladder/attribute/Height.java
@@ -1,4 +1,4 @@
-package domain.ladder.common;
+package domain.ladder.attribute;
 
 public class Height {
