+++ src/main/java/util/DirectionGenerator.java
@@ -1,6 +1,6 @@
 package util;
 
-import domain.ladder.common.Direction;
+import domain.ladder.attribute.Direction;
 
 public interface DirectionGenerator {
     Direction generate();
