+++ src/main/java/util/RandomDirectionGenerator.java
@@ -1,6 +1,6 @@
 package util;
 
-import domain.ladder.common.Direction;
+import domain.ladder.attribute.Direction;
 
 import java.util.Random;
