+++ src/test/java/domain/GameBoardTest.java
@@ -1,8 +1,8 @@
 package domain;
 
 import domain.ladder.Ladder;
-import domain.ladder.common.Direction;
-import domain.ladder.common.Height;
+import domain.ladder.attribute.Direction;
+import domain.ladder.attribute.Height;
 import domain.player.Name;
 import domain.player.Names;
 import domain.player.Player;
