+++ src/test/java/domain/PointTest.java
@@ -1,6 +1,6 @@
 package domain;
 
-import domain.ladder.common.Direction;
+import domain.ladder.attribute.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
